+++ service/conversations/v1/service/api_op_client.go
@@ -3,6 +3,8 @@ package service
 
 import (
 	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/binding"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/bindings"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/configuration"
 )
 
@@ -13,6 +15,8 @@ type Client struct {
 
 	sid string
 
+	Binding       func(string) *binding.Client
+	Bindings      *bindings.Client
 	Configuration func() *configuration.Client
 }
 
@@ -28,6 +32,15 @@ func New(client *client.Client, properties ClientProperties) *Client {
 
 		sid: properties.Sid,
 
+		Binding: func(bindingSid string) *binding.Client {
+			return binding.New(client, binding.ClientProperties{
+				ServiceSid: properties.Sid,
+				Sid:        bindingSid,
+			})
+		},
+		Bindings: bindings.New(client, bindings.ClientProperties{
+			ServiceSid: properties.Sid,
+		}),
 		Configuration: func() *configuration.Client {
 			return configuration.New(client, configuration.ClientProperties{
 				ServiceSid: properties.Sid,